moved functions from helpers.js into two new files: sequences.js and numbers.js
