+++ src/js/helpers.js
@@ -3,14 +3,10 @@
  */
 
 /* global alert, location, jQuery, localStorage, navigator */
-import { PRIMES, LINE_TYPE } from './constants.js'
+import { LINE_TYPE } from './constants.js'
 import { debug_enabled, resetTuningTable } from './scaleworkshop.js'
 import { SEMITONE_RATIO_IN_12_EDO } from './constants.js'
-
-// modulo function
-Number.prototype.mod = function (n) {
-  return ((this % n) + n) % n;
-};
+import { get_cf, get_convergent } from './sequences.js'
 
 // convert a cents value to decimal
 function cents_to_decimal(rawInput) {
@@ -214,96 +210,6 @@ function midi_note_number_to_name(input) {
   return name[remainder] + quotient;
 }
 
-// calculate the sum of the values in a given array given a stopping index
-function sum_array(array, index)
-{
-  var sum = 0;
-
-  if (array.length <= index)
-    index = array.length - 1;
-
-    for (let i = 0; i < index; i ++)
-    {
-      sum += array[i];
-    }
-
-    return sum;
-}
-
-// rotates the array by given steps
-function rotate(array, steps)
-{
-    var i = Math.abs(steps);
-    while (i > 0) {
-  var x;
-        if (steps < 0) {
-            x = array.shift();
-      array.push(x);
-  } else if (steps > 0) {
-      x = array.pop();
-      array.unshift(x);
-  }
-  i--;
-    }
-} 
-      
-// calculate a continued fraction for the given number
-function get_cf(num, maxiterations, roundf) {
-    var cf = [] // the continued fraction
-    var digit;
-    
-    var roundinv = 1.0 / roundf;
-    
-    var iterations = 0;
-    while (iterations < maxiterations)
-    {
-        digit = Math.floor(num);
-        cf.push(digit);
-        
-        num -= digit;
-        
-        if (num === 0 || num <= roundinv)
-        {
-            break;
-        }
-        
-        num = 1.0 / num;
-        iterations++;
-    }
-
-    return cf;
-}
-
-// calculate a single convergent for a given continued fraction
-function get_convergent(cf, depth=0) {
-
-    var cfdigit; // the continued fraction digit
-    var num; // the convergent numerator
-    var den; // the convergent denominator
-    var tmp; // for easy reciprocation
-
-  if (depth >= cf.length || depth === 0)
-    depth = cf.length;
-    
-    for (let d = 0; d < depth; d++)
-    {
-        cfdigit = cf[d];
-        num = cfdigit;
-        den = 1;
-        
-        // calculate the convergent
-        for (let i = d; i > 0; i--)
-        {
-            tmp = den;
-            den = num;
-            num = tmp;
-            num += den * cf[i - 1];
-        }
-    }
-
-  return num + '/' + den;
-}
-
 // convert a decimal to ratio (string 'x/y'), may have rounding errors for irrationals
 function decimal_to_ratio(rawInput, iterations=15, depth=0) {
 
@@ -332,467 +238,6 @@ function n_of_edo_to_ratio(rawInput, iterations=15, depth=0) {
   return decimal_to_ratio(n_of_edo_to_decimal(rawInput), iterations, depth);
 }
 */
-
-// calculate all best rational approximations given a continued fraction
-function get_convergents(cf, numarray, denarray, perlimit, cindOut=null)
-{
-    var cfdigit; // the continued fraction digit
-    var num; // the convergent numerator
-    var den; // the convergent denominator
-    var scnum; // the semiconvergent numerator
-    var scden; // the semiconvergen denominator
-    var cind = []; // tracks indicies of convergents
-    
-    for (let d = 0; d < cf.length; d++)
-    {
-        cfdigit = cf[d];
-        num = cfdigit;
-        den = 1;
-        
-        // calculate the convergent
-        for (let i = d; i > 0; i--)
-        {
-          [den, num] = [num, den]
-          num += den * cf[i - 1];
-        }
-
-        if (d > 0)
-        {
-          for (let i = 1; i < cfdigit; i++)
-          {
-              scnum = num - (cfdigit - i) * numarray[cind[d-1]];
-              scden = den - (cfdigit - i) * denarray[cind[d-1]];
-
-              if (scden > perlimit)
-                break;
-                
-              numarray.push(scnum);
-              denarray.push(scden);
-          }
-        }
-
-        if (den > perlimit)
-           break;
-                
-        cind.push(numarray.length);
-        numarray.push(num);
-        denarray.push(den);
-    }
-
-  if (!(cindOut===null)) 
-  {
-    for (let i = 0; i < cind.length; i++)
-    {
-      cindOut.push(cind[i]);
-    }
-  }
-
-    //for (let i = 0; i < denarray.length; i++)
-    //  console.log(numarray[i]+"/"+denarray[i]);
-}
-                         
-// generate and display MOS list
-function show_mos_cf(per, gen, ssz, threshold) {
-    var maxsize = 400; // maximum period size
-    var maxcfsize = 12; // maximum continued fraction length
-    var roundf = 1000; // rounding factor in case continued fraction blows up
-     
-    per = line_to_decimal(per);
-    if (per <= 0 || isNaN(per)) {
-    jQuery("#info_rank_2_mos").text("invalid period");
-    return false;
-    }
-
-    gen = line_to_decimal(gen);
-    if (gen <= 0 || isNaN(gen)) {
-    jQuery("#info_rank_2_mos").text("invalid generator");
-    return false;
-    }
-
-    var genlog = Math.log(gen) / Math.log(per); // the logarithmic ratio to generate MOS info
-
-    var cf = []; // continued fraction
-    var nn = []; // MOS generators
-    var dd = []; // MOS periods
-
-    cf = get_cf(genlog, maxcfsize, roundf);
-    get_convergents(cf, nn, dd, maxsize);
-    
-    // filter by step size threshold
-    var gc = decimal_to_cents(gen);
-    var pc = decimal_to_cents(per);
-    var L = pc + gc; // Large step
-    var s = pc; // small step
-    var c = gc; // chroma (L - s)
-
-    for (let i = 1; i < cf.length; i++)
-    {
-        L -= c * cf[i];
-        s = c;
-        c = L - s;
-        
-        // break if g is some equal division of period
-        if (c < (1 / roundf) && cf.length < maxcfsize)
-        {
-          // add size-1 
-          // not sure if flaw in the algorithm or weird edge case
-          
-          if (dd[dd.length-2] !== dd[dd.length-1]-1)
-            dd.splice(dd.length-1, 0, dd[dd.length-1]-1);
-
-          break;
-        }
-        
-        if (c < threshold)
-        {
-            var ind = sum_array(cf, i+1);
-            dd.splice(ind+1, dd.length - ind);
-            break;
-        }
-    }
-
-    // the first two periods are trivial
-    dd.shift();
-    dd.shift();
-     
-    jQuery("#info_rank_2_mos").text(dd.join(", "));
-}
-                 
-// helper function to simply pass in an interval and get an array of ratios returned
-function get_rational_approximations(intervalIn, numerators, denominators, roundf=999999,
-                                     cidxOut=null, ratiosOut=null, numlimits=null, denlimits=null, ratiolimits=null) {
-    
-    var cf = []; // continued fraction
-
-    cf = get_cf(intervalIn, 15, roundf);
-    get_convergents(cf, numerators, denominators, roundf, cidxOut);
-    
-    var doRatios = !(ratiosOut===null);
-    var doNumLim = !(numlimits===null);
-    var doDenLim = !(denlimits===null);
-    var doRatioLim = !(ratiolimits===null);
-                 
-    if (doRatios|| doNumLim || doDenLim || doRatioLim) {
-        var nlim;
-        var dlim;
-
-        for (let i = 0; i < numerators.length; i++) {
-            numerators[i] === 1 ? nlim = 1 : nlim = get_prime_limit(numerators[i]);
-            denominators[i] === 1 ? dlim = 1 : dlim = get_prime_limit(denominators[i]);
-
-            if (doRatios)
-                ratiosOut.push(numerators[i]+"/"+denominators[i]);
-            if (doNumLim)
-                numlimits.push(nlim);
-            if (doDenLim)
-                denlimits.push(dlim);
-            if (doRatioLim)
-                ratiolimits.push(Math.max(nlim, dlim));
-        }
-    }
-}
-
-// rank2 scale algorithm intended for integers, in ET contexts
-// for example, period = 12, gen = 7 : [ 2 2 1 2 2 2 1 ]
-function get_rank2_mode(period, generator, size, numdown=0) {
-  let degrees = [];
-  let modeOut = [];
-  var interval;
-
-  interval = generator * -numdown;
-  for (let n = 0; n < size; n++) {
-    while (interval < 0) {
-      interval += period;
-    }
-    if (interval >= period) {
-      interval %= period;
-    }
-    degrees.push(interval);
-    interval += generator;
-  }
-
-  degrees.sort(function(a, b) { return a-b });
-  for (let n = 1; n < degrees.length; n++) {
-    modeOut.push(degrees[n] - degrees[n-1]);
-  }
-
-  modeOut.push(period - degrees[degrees.length-1]);
-
-  return modeOut;
-}
-
-// returns an array representing the prime factorization
-// indicies are the 'nth' prime, the value is the powers of each prime
-function get_prime_factors(number) {
-    number = Math.floor(number);
-    if (number === 1) {
-        //alert("Warning: 1 has no prime factorization.");
-        return 1;
-     }
-    var factorsout = [];
-    var n = number;
-    var q = number;
-    var loop;
-                 
-    for (let i = 0; i < PRIMES.length; i++) {
-        if (PRIMES[i] > n)
-            break;
-                 
-         factorsout.push(0);
-                 
-        if (PRIMES[i] === n) {
-            factorsout[i]++;
-            break;
-        }
-                 
-        loop = true;
-         
-        while (loop) {
-            q = n / PRIMES[i];
-        
-            if (q === Math.floor(q)) {
-                n = q;
-                factorsout[i]++;
-                continue;
-            }
-            loop = false;
-         }
-     }
-    
-    return factorsout;
-}
-
-/*
-function get_prime_factors_string(number) {
-     var factors = get_prime_factors(number);
-     var str_out = "";
-                 
-     for (let i = 0; i < factors.length; i++) {
-                 
-         if (factors[i] !== 0) {
-            str_out += PRIMES[i] + "^" + factors[i];
-                 
-            if (i < factors.length - 1)
-                str_out += " * ";
-         }
-     }
-    return str_out;
- }
- */
-
- function isPrime(number) {
-    var sqrtnum = Math.floor(Math.sqrt(number));
-    
-    for (let i = 0; i < PRIMES.length; i++)
-    {
-        if (PRIMES[i] >= sqrtnum)
-            break;
-    
-        if (number % PRIMES[i] === 0) {
-            return false;
-        }
-    }
-    return true;
- }
-                 
-function prevPrime(number)
-{
-  if (number < 2)
-    return 2;
-    var i = 0;
-    while (i < PRIMES.length && PRIMES[i++] <= number);
-    return PRIMES[i - 2];
-}
-                 
-function nextPrime(number)
-{
-  if (number < 2)
-    return 2;
-     var i = 0;
-     while (i < PRIMES.length && PRIMES[i++] <= number);
-     return PRIMES[i - 1];
-}
-
-function closestPrime(number)
-{
-  var thisPrime = isPrime(number);
-  
-  if (number < 2)
-    return 2;
-  else if (thisPrime)
-    return number;
-
-  var np = nextPrime(number);
-  var pp = prevPrime(number);
-
-  if (Math.abs(np - number) < Math.abs(pp - number))
-    return np;
-  else
-    return pp;
-}
-
-/*
-function scrollToPrime(number, scrollDown)
-{
-    if (scrollDown)
-        return prevPrime(number);
-    else
-        return nextPrime(number);
-}
-*/
-
-function get_prime_limit(number) {
-  var factors = get_prime_factors(number);
-  return PRIMES[factors.length - 1];
-}
-
-/*
-function get_prime_limit_of_ratio(numerator, denominator) {
-  return Math.max(get_prime_limit(numerator), get_prime_limit(denominator));
-}
-*/
-
- // returns an array of integers that share no common factors to the given integer
- function get_coprimes(number) {
-    let coprimes = [1];
-    var m, d, t;
-    for (let i = 2; i < number - 1; i++) {
-        m = number;
-        d = i;
-        while (d > 1) {
-            m = m % d;
-            t = d;
-            d = m;
-            m = t;
-        }
-        if (d > 0) {
-            coprimes.push(i);
-        }
-    }
-    coprimes.push(number-1);
-    return coprimes;
- }
-
-/*
-// returns an array of integers that can divide evenly into given number
-function get_factors(number) {
-  let factors = [];
-  const nsqrt = Math.floor(Math.sqrt(number));
-
-  for (let n = 2; n <= nsqrt; n++) {
-    let q = number / n;
-    if (Math.floor(q) === q) {
-      factors.push(n);
-      if (n !== q)
-        factors.push(q);
-    }
-  }
-
-  return factors.sort(function(a, b) { return a-b });;
-}
-*/
-
- // returns array of the numerator and denominator of the reduced form of given ratio
- function reduce_ratio(numerator, denominator) { 
-  var num_pf = get_prime_factors(numerator);
-  var den_pf = get_prime_factors(denominator);
-  let r_pf = [];
-  var maxlength = Math.max(num_pf.length, den_pf.length);
-  for(var i = 0; i < maxlength; i++) {
-    var sum = 0;
-
-    if (i < num_pf.length) {
-      sum = num_pf[i];
-    }
-
-    if (i < den_pf.length) {
-      sum -= den_pf[i];
-    }
-
-    r_pf.push(sum);
-  }
-
-  var nn = 1;
-  var dd = 1;
-
-  for (let i = 0; i < maxlength; i++) {
-    if (r_pf[i] > 0)
-      nn *= Math.pow(PRIMES[i], r_pf[i]);
-    else
-      dd *= Math.pow(PRIMES[i], r_pf[i] * -1);
-  }
-
-  return [nn, dd];
- }
-
- function get_lcm(array) {
-    let primecounts = [];
-   let primefactors = [];
-   var f;
-   array.forEach(function(item, index, array) {
-    f = get_prime_factors(item);
-    primefactors.push(f);
-   });
-   
-   var maxlength = 0;
-   primefactors.forEach(function(item, index, array) {
-    if (item.length > maxlength)
-      maxlength = item.length;
-   });
-
-   // find the min power of each primes in numbers' factorization
-   for (let p = 0; p < maxlength; p++) {
-    primecounts.push(0);
-     for (let n = 0; n < primefactors.length; n++) {
-      f = primefactors[n];
-      if (p < f.length) {
-        if (primecounts[p] < f[p])
-          primecounts[p] = f[p];
-      }
-     }	 
-   }
-
-   let lcm = 1;
-   primecounts.forEach(function(item, index) {
-    lcm *= Math.pow(PRIMES[index], item);
-   });
-
-   return lcm;
-}
-
- function invert_chord(chord) {
-  if (!/^(\d+:)+\d+$/.test(chord)) {
-    alert("Warning: invalid chord " + chord);
-    return false;
-  }
-
-  let inverted = chord;
-  let intervals = chord.split(":").map(x => parseInt(x));
-  let steps = [];
-  intervals.forEach(function(item, index, array) {
-    if (index > 0) {
-      steps.push([item, array[index-1]]);
-    }
-  })
-  steps.reverse();
-  intervals = [[1, 1]];
-  
-  let denominators = [];
-  steps.forEach(function(item, index) {
-    var reduced_interval = reduce_ratio(item[0] * intervals[index][0], item[1] * intervals[index][1]);
-    intervals.push(reduced_interval);
-    denominators.push(reduced_interval[1]);
-  });
-  
-  var lcm = get_lcm(denominators);
-
-  chord = [];
-  intervals.forEach(function(x) {
-    chord.push(x[0] * lcm / x[1]);
-  });
-
-  return chord.join(":");
- }
 
 function debug(msg = "") {
   if (debug_enabled) {
@@ -959,15 +404,9 @@ export {
   setScaleName,
   getFloat,
   getString,
-  invert_chord,
   mtof,
   midi_note_number_to_name,
   ftom,
-  get_coprimes,
-  get_rank2_mode,
-  get_rational_approximations,
-  rotate,
-  closestPrime,
   isFunction,
   line_to_cents,
   openDialog,
@@ -976,5 +415,4 @@ export {
   n_of_edo_to_cents,
   isLocalStorageAvailable,
   isRunningOnWindows,
-  show_mos_cf
 }
